libghdl*.so is now part of GHDL
* move python sources to python/libghdl and python/pnodes
* rename src/vhdl/python to src/vhdl/libghdl
* add generation of tarball for libghdl-py to the makefile
* deprecate --enable-python and --disable-python
* add configuration option --disable-libghdl
