--- libghdl_pkg.py.in
@@ -1,32 +1,0 @@
-#!/usr/bin/env python
-
-from os import mkdir
-from os.path import join
-from shutil import copy, copytree
-
-
-libghdl_filename = 'libghdl-' + "@libghdl_version@" + '@SOEXT@'
-dstdir = 'build-libghdl'
-
-# Copy python sources
-copytree(
-    join('@srcdir@/src/vhdl/python/libghdl'),
-    join(dstdir),
-)
-
-# Create config file
-with open(join(dstdir, 'config.py'), 'w') as f:
-    f.write('libghdl_filename="{}"\n'.format(libghdl_filename))
-
-# Copy shared library and VHDL libraries
-copy(libghdl_filename, dstdir)
-copy('setup.py', dstdir)
-
-# Copy VHDL libraries
-def ignore_function(_, lst):
-    return [item for item in lst if item.endswith('.o')]
-copytree(
-    join('lib', 'ghdl'),
-    join(dstdir, 'ghdl'),
-    ignore=ignore_function
-)
