+++ python/libghdl/thin/errorout.py
