+++ python/libghdl/thin/errorout_memory.py
