+++ python/libghdl/thin/files_map.py
