+++ python/libghdl/thin/files_map_editor.py
