+++ python/libghdl/thin/flags.py
