+++ python/libghdl/thin/libraries.py
