+++ python/libghdl/thin/name_table.py
