+++ python/libghdl/thin/std_names.py
