+++ python/libghdl/thin/vhdl/canon.py
