+++ python/libghdl/thin/vhdl/elocations.py
