+++ python/libghdl/thin/vhdl/ieee.py
