+++ python/libghdl/thin/vhdl/nodes.py
