+++ python/libghdl/thin/vhdl/nodes_meta.py
