+++ python/libghdl/thin/vhdl/nodes_utils.py
