+++ python/libghdl/thin/vhdl/parse.py
