+++ python/libghdl/thin/vhdl/pyutils.py
