+++ python/libghdl/thin/vhdl/scanner.py
