+++ python/libghdl/thin/vhdl/sem.py
