+++ python/libghdl/thin/vhdl/sem_lib.py
