+++ python/libghdl/thin/vhdl/tokens.py
